Rust コマンドを clap 化し Cargo workspace で独立バージョニング + E2E テスト化 (#406) (#407)
## 概要

Epic #388 の次段。ルート単一クレートを **root package を持つ Cargo workspace**
へ再編し、各配布コマンドを独立クレート・独立バージョニング（release-plz workspace）にし、E2E テストを備える。Refs
#406 / #388。

> マージ責任はユーザー（しっかりレビューのうえ Epic へマージ）。Epic→main は #403 で別途。

## 変更内容

### 1. Cargo workspace 化（`refactor`）
- ルート `Cargo.toml` を **root package `dotfiles`(0.68.0継続) +
`[workspace]`** に。`[workspace.package/dependencies/lints]` で共通メタ・依存・lint
を集約。
- `src/lib.rs` → `crates/dotfiles-support`（`command_exists`, 0.1.0 /
publish=false / 非install）。
- 衛星 → `crates/{color,copy-obj,v-sync,gh-clone,gcm,git-upstream}`（各
0.1.0）。
- 背景 worker → `crates/dotfiles-workers`（2 bin, 0.1.0）。
- lint → `tools/dotfiles-lint`（0.0.0・版なし・非install・publish=false）。
- 新規 `src/main.rs`（root `dotfiles` bin: clap `--version`/`--help`）。`git
mv` で履歴保持。

### 2. clap 化（`feat`）
- color/copy-obj/v-sync/gh-clone/gcm に clap derive を追加し手書き args
パースを廃止。git-upstream は `version` 追加。worker は env 駆動のため非 clap。
- 全配布 bin（worker 除く）が `--help`/`--version`（`CARGO_PKG_VERSION`）対応。

### 3. E2E（`test`）
- root + 衛星6 に `tests/e2e/mod.rs`（`assert_cmd` + `predicates` +
`rstest`）。`PATH=/nonexistent` で依存不在時の exit code（127 / 8）や必須引数欠落（exit
2）を決定的に検証。
- support は unit、workers は最小 smoke。

### 4. release-plz（`ci`）
- workspace・per-package。root は `v{version}` 維持、member は
`<crate>-v{version}`、`dotfiles-lint` は `release = false`。

### 5–6. CI / chezmoi フック（`ci` / `chore`）
- rust.yml を `--workspace` 化、apply-test で配布物 install と lint/support
非install を検証。
- cargo-install フックを配布クレートのみ共有 `--target-dir` で順次 install するループへ。

### 7. ドキュメント（`docs` / `style`）
- README / CONTRIBUTING / CLAUDE / mise.toml / lint.yml / fish コメントを
workspace 構成へ。

## 検証（ローカル, すべて green）

- `cargo fmt --all --check` / `cargo clippy --workspace --all-targets
--locked -- -D warnings`
- `cargo build --workspace --locked` / `cargo test --workspace --locked`
→ **58 passed**
- `mise run check`（dotfiles-lint）→ failures=0
- 一時 root へ `cargo install` → 配布 9 bin（`dotfiles --version` = `dotfiles
0.68.0`）installed、`dotfiles-lint`/`dotfiles-support` は非install を確認

## 完了条件（Issue #406）

- [x] root `dotfiles` + 衛星6 + support + workers が各独立クレート、lint は `tools/`
の非配布 member
- [x] 配布物・lib は独立版（root v{version} 継続、新規 0.1.0）、lint は版なし
- [x] worker 以外の bin が clap（`--help`/`--version`）
- [x] 各配布 crate に E2E（assert_cmd + predicates + rstest）
- [x] release-plz workspace で per-package
タグ（root=`v{version}`、member=`<crate>-v<ver>`）
- [x] `cargo install`（chezmoi フック）で配布物が `~/.cargo/bin`（lint は入らない）
- [ ] CI 全 green（fmt/clippy/test/E2E/lint/apply-test）← この PR の CI で確認
- [ ] `epic/388-rust-cargo-release` へマージ（その後 #403 を再確認）← ユーザー


---------
