+++ CONTRIBUTING.md
@@ -33,7 +33,9 @@ dotfiles の利用に必要なツールは [`README.md`](README.md) を参照し
 
 ## リリースプロセス
 
-このリポジトリは [release-plz](https://release-plz.dev/) を使用して、Conventional Commits に基づくバージョン管理を自動化しています。version の source of truth は `Cargo.toml`、CHANGELOG は git-cliff（`cliff.toml`）で生成します。crates.io へは publish せず、git タグ + GitHub Release のみを作成します。
+このリポジトリは [release-plz](https://release-plz.dev/) を使用して、Conventional Commits に基づくバージョン管理を自動化しています。version の source of truth は各 `Cargo.toml`、CHANGELOG は git-cliff（`cliff.toml`）で生成します。crates.io へは publish せず、git タグ + GitHub Release のみを作成します。
+
+**Cargo workspace の per-package バージョニング**: 配布物（root `dotfiles` + 各コマンド + workers）と共有 lib（`dotfiles-support`）はそれぞれ独立に版を振ります。タグは root `dotfiles` が `v{version}`（従来どおり）、その他の member が `<crate>-v{version}`。CHANGELOG は各クレート直下に per-package で生成されます。開発専用ツール `tools/dotfiles-lint` は `release = false` で release-plz の対象外（版を振りません）。下表のバンプ規則はパッケージごとに、そのパッケージに触れたコミットの type で判定されます。
 
 ### バージョン決定ルール
 
@@ -48,14 +50,17 @@ dotfiles の利用に必要なツールは [`README.md`](README.md) を参照し
 
 1. `mise run release-prepare`（または GitHub の Release Prepare workflow を `workflow_dispatch`）で release-plz が Release PR（`release-*` ブランチ）を作成・更新する。
 2. Release PR には `Cargo.toml` の version と `CHANGELOG.md` の更新が含まれる（必要なら PR 上で version を手動調整できる）。
-3. Release PR を `main` にマージすると Release Publish workflow が走り、`v{version}` タグと GitHub Release が作成される。
+3. Release PR を `main` にマージすると Release Publish workflow が走り、per-package タグ（root は `v{version}`、その他は `<crate>-v{version}`）と GitHub Release が作成される。
 
 `main` への直接 push は不要・不可。Release PR のマージがリリースのトリガー（ブランチ保護と両立）。
 
 ### 現在のバージョン確認
 
 ```bash
-git describe --tags --abbrev=0
+# root dotfiles（core）
+git describe --tags --abbrev=0 --match 'v[0-9]*'
+# 各コマンド（例: color）
+git describe --tags --abbrev=0 --match 'color-v*'
 ```
 
 ## テスト
@@ -72,19 +77,19 @@ mise run lint
 mise run check
 ```
 
-lint は Rust 製オーケストレータ `dotfiles-lint`（`src/bin/dotfiles-lint`）が、mise 供給の
-各ツールを呼んで実行します。オーケストレータ自体を変更したときは、以下でロジックの
-ユニットテストを確認してください:
+lint は Rust 製オーケストレータ `dotfiles-lint`（`tools/dotfiles-lint`）が、mise 供給の
+各ツールを呼んで実行します。コマンドや lint オーケストレータを変更したときは、以下で
+workspace 全体のテスト（各クレートの unit + E2E）を確認してください:
 
 ```bash
-cargo test
+cargo test --workspace
 ```
 
 補足:
 
-- CI は `cargo run --bin dotfiles-lint -- check`（lint）と `cargo test`（オーケストレータの
-  テスト）を実行します。
-- 詳細ログが必要な場合は `cargo run --bin dotfiles-lint -- check --summary --json` を使ってください。
+- CI は `cargo run -p dotfiles-lint -- check`（lint）と `cargo test --workspace`（unit + E2E）を
+  実行します。
+- 詳細ログが必要な場合は `cargo run -p dotfiles-lint -- check --summary --json` を使ってください。
 
 ## コードスタイル
