+++ crates/color/Cargo.toml
@@ -1,0 +1,24 @@
+# ANSI カラー表を出力する配布コマンド。独立版・独自タグ `color-v<ver>`。
+[package]
+description = "ANSI カラーコード（16 色 + 256 色）の確認表を出力する"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "color"
+publish     = false
+repository  = { workspace = true }
+version     = "0.1.0"
+
+[dependencies]
+clap = { workspace = true }
+
+[dev-dependencies]
+assert_cmd = { workspace = true }
+predicates = { workspace = true }
+rstest     = { workspace = true }
+
+[[test]]
+name = "e2e"
+path = "tests/e2e/mod.rs"
+
+[lints]
+workspace = true
