+++ crates/dotfiles-workers/Cargo.toml
@@ -1,0 +1,24 @@
+# バックグラウンド worker 2 種（daily-check-worker / git-background-fetch-worker）の配布クレート。
+# fish conf.d のフックからバックグラウンドで起動される。clap 化はしない（引数を取らず env 駆動）。
+[package]
+description = "dotfiles のバックグラウンド worker（daily-check / git-background-fetch）"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "dotfiles-workers"
+publish     = false
+repository  = { workspace = true }
+version     = "0.1.0"
+
+[dependencies]
+dotfiles-support = { path = "../dotfiles-support" }
+
+[dev-dependencies]
+assert_cmd = { workspace = true }
+predicates = { workspace = true }
+
+[[test]]
+name = "e2e"
+path = "tests/e2e/mod.rs"
+
+[lints]
+workspace = true
