+++ crates/gcm/Cargo.toml
@@ -1,0 +1,26 @@
+# ステージ済み diff から Conventional Commits を AI 生成して commit する配布コマンド。
+[package]
+description = "AI-powered git commit with Conventional Commits（claude -p）"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "gcm"
+publish     = false
+repository  = { workspace = true }
+version     = "0.1.0"
+
+[dependencies]
+clap       = { workspace = true }
+serde      = { workspace = true }
+serde_json = { workspace = true }
+
+[dev-dependencies]
+assert_cmd = { workspace = true }
+predicates = { workspace = true }
+rstest     = { workspace = true }
+
+[[test]]
+name = "e2e"
+path = "tests/e2e/mod.rs"
+
+[lints]
+workspace = true
