+++ crates/git-upstream/Cargo.toml
@@ -1,0 +1,25 @@
+# upstream リモートを追従するマージヘルパー配布コマンド（git upstream サブコマンド）。
+[package]
+description = "upstream/master を fetch してマージする / upstream リモートを初期化する"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "git-upstream"
+publish     = false
+repository  = { workspace = true }
+version     = "0.1.0"
+
+[dependencies]
+clap = { workspace = true }
+libc = { workspace = true }
+
+[dev-dependencies]
+assert_cmd = { workspace = true }
+predicates = { workspace = true }
+rstest     = { workspace = true }
+
+[[test]]
+name = "e2e"
+path = "tests/e2e/mod.rs"
+
+[lints]
+workspace = true
