+++ crates/v-sync/Cargo.toml
@@ -1,0 +1,25 @@
+# nvim プラグインを同期し lazy-lock.json を chezmoi ソースへ re-add する配布コマンド。
+[package]
+description = "Neovim プラグインを同期し lazy-lock.json を chezmoi ソースへ re-add する"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "v-sync"
+publish     = false
+repository  = { workspace = true }
+version     = "0.1.0"
+
+[dependencies]
+clap             = { workspace = true }
+dotfiles-support = { path = "../dotfiles-support" }
+
+[dev-dependencies]
+assert_cmd = { workspace = true }
+predicates = { workspace = true }
+rstest     = { workspace = true }
+
+[[test]]
+name = "e2e"
+path = "tests/e2e/mod.rs"
+
+[lints]
+workspace = true
