+++ home/.chezmoiremove.tmpl
@@ -7,11 +7,11 @@
 {{/* fzf fish 関数のリネーム (_fzf_worktree → _fzf_worktree_remove / _fzf_ghq_repo → _fzf_ghq_cd) に伴う旧ファイル削除 */}}
 .config/fish/functions/_fzf_worktree.fish
 .config/fish/functions/_fzf_ghq_repo.fish
-{{/* gcm を Rust バイナリ化 (src/bin/gcm, cargo install で ~/.cargo/bin) したため fish 関数を削除 */}}
+{{/* gcm を Rust バイナリ化 (crates/gcm, cargo install で ~/.cargo/bin) したため fish 関数を削除 */}}
 .config/fish/functions/gcm.fish
-{{/* copy-obj / v-sync を Rust バイナリ化 (src/bin/, cargo install) したため fish 関数を削除 */}}
+{{/* copy-obj / v-sync を Rust バイナリ化 (crates/copy-obj, crates/v-sync, cargo install) したため fish 関数を削除 */}}
 .config/fish/functions/copy-obj.fish
 .config/fish/functions/v-sync.fish
-{{/* 背景 worker を Rust バイナリ化 (src/bin/daily-check-worker, git-background-fetch-worker) したため fish 関数を削除 */}}
+{{/* バックグラウンド worker を Rust バイナリ化 (crates/dotfiles-workers: daily-check-worker, git-background-fetch-worker) したため fish 関数を削除 */}}
 .config/fish/functions/_daily_check_worker.fish
 .config/fish/functions/_git_background_fetch_worker.fish
