+++ home/dot_config/fish/functions/gh-clone.fish
@@ -1,4 +1,4 @@
-# gh-clone のロジックは Rust バイナリ（src/bin/gh-clone, ~/.cargo/bin）に移した。
+# gh-clone のロジックは Rust バイナリ（crates/gh-clone, ~/.cargo/bin）に移した。
 # 別プロセスでは親シェルの cd を変えられないため、ここはバイナリが出力した
 # 移設先パスへ cd するだけの薄い shim（zoxide / starship と同じ定石）。
 function gh-clone --description 'Clone repository via gh and migrate with ghq'
