+++ tests/e2e/mod.rs
@@ -1,0 +1,37 @@
+//! root `dotfiles`（core）bin の E2E テスト（assert_cmd + predicates + rstest）。
+
+use assert_cmd::Command;
+use predicates::prelude::*;
+use rstest::rstest;
+
+fn dotfiles() -> Command {
+    Command::cargo_bin("dotfiles").unwrap()
+}
+
+#[rstest]
+#[case("--help")]
+#[case("-h")]
+fn help_flag_succeeds(#[case] flag: &str) {
+    dotfiles().arg(flag).assert().success();
+}
+
+#[rstest]
+#[case("--version")]
+#[case("-V")]
+fn version_flag_prints_name_and_version(#[case] flag: &str) {
+    dotfiles()
+        .arg(flag)
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("dotfiles"))
+        .stdout(predicate::str::contains(env!("CARGO_PKG_VERSION")));
+}
+
+#[test]
+fn no_args_prints_version() {
+    dotfiles()
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("dotfiles"))
+        .stdout(predicate::str::contains(env!("CARGO_PKG_VERSION")));
+}
