+++ tools/dotfiles-lint/Cargo.toml
@@ -1,0 +1,18 @@
+# lint/format オーケストレータ。repo 内部の開発専用ツール（xtask 定石）。
+# 版を振らず（release-plz は release=false）、install もしない。publish もしない。
+[package]
+description = "dotfiles の lint/format オーケストレータ（repo 内部の開発専用ツール）"
+edition     = { workspace = true }
+license     = { workspace = true }
+name        = "dotfiles-lint"
+publish     = false
+repository  = { workspace = true }
+version     = "0.0.0"
+
+[dependencies]
+clap     = { workspace = true }
+ignore   = { workspace = true }
+tempfile = { workspace = true }
+
+[lints]
+workspace = true
