+++ tools/dotfiles-lint/src/lint.rs
