+++ tools/dotfiles-lint/src/lint/classify.rs
