+++ tools/dotfiles-lint/src/lint/collect.rs
