+++ tools/dotfiles-lint/src/main.rs
